+++ ChangeLog.md
@@ -3,6 +3,13 @@ MongoDB for XP Framework ChangeLog
 
 ## ?.?.? / ????-??-??
 
+## 0.4.0 / 2020-08-07
+
+* Changed `insert()` to generate and return object IDs if not supplied in
+  passed documents. Consistent with how most MongoDB drivers handle this, see
+  https://docs.mongodb.com/manual/reference/method/db.collection.insert/#id-field
+  (@thekid)
+
 ## 0.3.1 / 2020-08-06
 
 * Fixed large decimal128 number parsing from a string - @thekid